Extract header and footer into containers
