+++ src/App.css
@@ -16,75 +16,3 @@ a {
 a:hover {
     color: var(--font-color-secondary);
 }
-
-.header {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-.footer {
-    position: relative;
-    top: 3000px;
-    font-family: var(--font-primary);
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.footer__left {
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-}
-
-.footer__right {
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-}
-
-.footer__right__title {
-    font-size: 24px;
-    font-weight: bold;
-    color: var(--font-color-secondary);
-}
-
-.footer__right__links {
-    font-size: 18px;
-    display: flex;
-    gap: 24px;
-}
-
-/* Responsiveness */
-
-@media screen and (min-width: 1200px) {
-    .header, .footer {
-        padding: 32px 48px;
-    }
-    .footer__right__links {
-        flex-direction: row;
-    }
-}
-
-@media screen and (max-width: 1200px) {
-    .header, .footer {
-        padding: 24px 32px;
-    }
-    .footer__right__links {
-        flex-direction: row;
-    }
-}
-
-@media screen and (max-width: 600px) {
-    .header, .footer {
-        padding: 16px 16px;
-    }
-    .footer__right__links {
-        flex-direction: column;
-    }
-    .footer {
-        align-items: baseline;
-    }
-}
