+++ src/containers/header/Header.css
@@ -1,0 +1,25 @@
+.header {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
+/* Responsiveness */
+/* TODO burger menu display:none*/
+@media screen and (min-width: 1200px) {
+    .header {
+        padding: 32px 48px;
+    }
+}
+
+@media screen and (max-width: 1200px) {
+    .header {
+        padding: 24px 32px;
+    }
+}
+
+@media screen and (max-width: 600px) {
+    .header {
+        padding: 16px 16px;
+    }
+}
